# Add .reserve(...) to assembleGeometry function